+++ Cargo.toml
@@ -76,3 +76,4 @@ tempfile = "3"
 insta = "1"
 wiremock = "0.6"
 rstest = "0.23"
+image = { version = "0.25", default-features = false, features = ["png"] }
