+++ src/instance/content/worlds.rs
@@ -135,79 +135,3 @@ fn format_size(bytes: u64) -> String {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn setup_saves_dir(tmp: &Path, instance: &str) -> std::path::PathBuf {
-        let dir = tmp.join(instance).join(".minecraft").join("saves");
-        std::fs::create_dir_all(&dir).unwrap();
-        dir
-    }
-
-    #[test]
-    fn scan_worlds_empty_dir() {
-        let tmp = tempfile::tempdir().unwrap();
-        setup_saves_dir(tmp.path(), "inst");
-        let worlds = scan_worlds(tmp.path(), "inst");
-        assert!(worlds.is_empty());
-    }
-
-    #[test]
-    fn scan_worlds_missing_dir_returns_empty() {
-        let tmp = tempfile::tempdir().unwrap();
-        let worlds = scan_worlds(tmp.path(), "ghost");
-        assert!(worlds.is_empty());
-    }
-
-    #[test]
-    fn scan_worlds_finds_directories() {
-        let tmp = tempfile::tempdir().unwrap();
-        let dir = setup_saves_dir(tmp.path(), "inst");
-        std::fs::create_dir(dir.join("My World")).unwrap();
-        std::fs::create_dir(dir.join("Creative")).unwrap();
-        let worlds = scan_worlds(tmp.path(), "inst");
-        assert_eq!(worlds.len(), 2);
-    }
-
-    #[test]
-    fn scan_worlds_ignores_files() {
-        let tmp = tempfile::tempdir().unwrap();
-        let dir = setup_saves_dir(tmp.path(), "inst");
-        std::fs::create_dir(dir.join("World1")).unwrap();
-        std::fs::write(dir.join("stray-file.txt"), "not a world").unwrap();
-        let worlds = scan_worlds(tmp.path(), "inst");
-        assert_eq!(worlds.len(), 1);
-    }
-
-    #[test]
-    fn scan_worlds_disabled_world() {
-        let tmp = tempfile::tempdir().unwrap();
-        let dir = setup_saves_dir(tmp.path(), "inst");
-        std::fs::create_dir(dir.join("ActiveWorld")).unwrap();
-        std::fs::create_dir(dir.join("HiddenWorld.disabled")).unwrap();
-        let worlds = scan_worlds(tmp.path(), "inst");
-        let active = worlds
-            .iter()
-            .find(|w| w.file_stem == "ActiveWorld")
-            .unwrap();
-        let hidden = worlds
-            .iter()
-            .find(|w| w.file_stem == "HiddenWorld")
-            .unwrap();
-        assert!(active.enabled);
-        assert!(!hidden.enabled);
-    }
-
-    #[test]
-    fn scan_worlds_sorted_case_insensitive() {
-        let tmp = tempfile::tempdir().unwrap();
-        let dir = setup_saves_dir(tmp.path(), "inst");
-        std::fs::create_dir(dir.join("Zeta")).unwrap();
-        std::fs::create_dir(dir.join("alpha")).unwrap();
-        std::fs::create_dir(dir.join("Beta")).unwrap();
-        let worlds = scan_worlds(tmp.path(), "inst");
-        let names: Vec<&str> = worlds.iter().map(|w| w.name.as_str()).collect();
-        assert_eq!(names, vec!["alpha", "Beta", "Zeta"]);
-    }
-}
